Changed mac from int to long long

+++ QTUI/chartView.h
@@ -18,7 +18,7 @@ class chartview : public QChartView
 public:
     chartview(QWidget *parent = 0);
     ~chartview();
-    std::list<std::pair<float, int>> probableHiddenMatching(position, std::vector<position>);
+    std::list<std::pair<float, long long>> probableHiddenMatching(position, std::vector<position>);
 
 public Q_SLOTS:
     void handleClickedPoint(const QPointF &point);
